fix(header): refactor Header
