--- src/components/Header.astro
@@ -1,7 +1,0 @@
----
-import { HeaderTransition } from "./HeaderTransition";
----
-
-<div class="flex justify-center items-center">
-  <HeaderTransition client:load />
-</div>
